Задание 5. Часть 1: Модуляция
