+++ js/variables.js
@@ -1,0 +1,55 @@
+const AMOUNT_PHOTOS_ID = {
+  MIN: 1,
+  MAX: 25
+};
+
+const DESCRIPTIONS = [
+  'Залив с песчанным пляжем',
+  'Дорожный указатель к пляжу',
+  'Камни у берега моря',
+  'Девушка в купальнике',
+  'Миски с рисом и супом',
+  'Чёрный спорткар',
+  'Клубника на тарелке',
+  'Фруктовый сок',
+  'Пролетающий самолёт над пляжем с отдыхающими',
+  'Обувная полка',
+  'Тропинка к пляжу, между оградой',
+  'Автомобиль ауди',
+  'Овощное блюдо с мясом',
+  'Суши-кот',
+  'Домашная обувь',
+  'Летящий самолет в небе',
+  'Хоровое пение',
+  'Автомобиль Кадиллак',
+  'Тапки с подсветкой',
+  'Аллея из пальм',
+  'Экзотическое блюдо',
+  'Закат у берега моря',
+  'Краб на камне',
+  'На зрительской трибуне большого концерта',
+  'Проезжабющий внедорожник брод с бегомотами'
+];
+
+const LIKES = {
+  MIN: 15,
+  MAX: 200
+};
+
+const AVATAR_NUMBER = {
+  MIN: 1,
+  MAX: 6
+};
+
+const MESSAGES = [
+  'Всё отлично!',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.',
+  'В целом всё неплохо.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+  'В конце концов это просто непрофессионально.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.'
+];
+
+const NAMES = ['Rachel', 'Monica', 'Phoebe', 'Joey', 'Chandler', 'Ross'];
+
+export {AMOUNT_PHOTOS_ID, DESCRIPTIONS, LIKES, AVATAR_NUMBER, MESSAGES, NAMES};
